services/backend-python-template/pyproject.toml
[project]
name = "backend-python-template"
version = "0.1.0"
description = "Python microservice template with JWT auth and OpenTelemetry"
authors = [
    {name = "Platform Team", email = "team@platform.com"},
]
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.1",
    "httpx==0.27.0",
    "python-jose[cryptography]==3.3.0",
    "prometheus-client==0.20.0",
    "pydantic==2.7.1",
    "python-multipart==0.0.9",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "httpx==0.27.0",
    "black==24.4.2",
    "isort==5.13.2",
    "flake8==7.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
